py-core/requirements.txt
# Python 2.7 Dependencies
# Note: Python 2.7 reached end of life in 2020

Flask==1.1.4
Werkzeug==1.0.1
psycopg2-binary==2.8.6
MarkupSafe==1.1.1
Jinja2==2.11.3
itsdangerous==1.1.0
click==7.1.2
